changed completed method and reworked edit to work with it
now when a task is completed, the date_completed gets changed from null to the current date. Added a check mark for the completed tasks instead of the 'X' that deletes the tasks. Created a function that checks if the task is complete and if so then the complete and edit buttons won't be displayed